I can't break the build ceiling!
I've tried all the obvious things, changing Y_SIZE to 514 (since it
was 256+2 before ...), changed everywhere I could find 256 related
to height/y checking to Y_SIZE-2. even paste() which doesn't
seem to be used anywhere that matters and I'm not quite sure
what the /copy /paste commands are supposed to do ...

My only thought is it might have something to do with the
underlying data structures (which might be why the limit
in the first place).